Fix ProxiedAction.
We don't actually need ProxiedCommand for this - AbstractCommand already
doesn't check CRCs. We do still need ProxiedCommand in order to get the
bytes stored somewhere sane - finalizeCmd does all the rest of the work.

There's still something wrong here because Club Caribe isn't actually
working - however, the traces clearly show that exactly the data that
should be making it to the C64 is making it there.